pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "celsius"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "U.N. Owen", email = "void@some.where" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "w1thermsensor",
  "rpi-lgpio",
  "flask",
]

[project.urls]
Documentation = "https://github.com/U.N. Owen/celsius#readme"
Issues = "https://github.com/U.N. Owen/celsius/issues"
Source = "https://github.com/U.N. Owen/celsius"

[tool.hatch.version]
path = "src/celsius/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/celsius tests}"

[tool.coverage.run]
source_pkgs = ["celsius", "tests"]
branch = true
parallel = true
omit = [
  "src/celsius/__about__.py",
]

[tool.coverage.paths]
celsius = ["src/celsius", "*/celsius/src/celsius"]
tests = ["tests", "*/celsius/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
